Add mutation, fix generic TableInput
