+++ graphql.schema.json
@@ -10713,6 +10713,41 @@
             "deprecationReason": null
           },
           {
+            "name": "insertValuationRules",
+            "description": "",
+            "args": [
+              {
+                "name": "request",
+                "description": "",
+                "type": {
+                  "kind": "INPUT_OBJECT",
+                  "name": "InsertValuationRulesInput",
+                  "ofType": null
+                },
+                "defaultValue": null
+              }
+            ],
+            "type": {
+              "kind": "NON_NULL",
+              "name": null,
+              "ofType": {
+                "kind": "LIST",
+                "name": null,
+                "ofType": {
+                  "kind": "NON_NULL",
+                  "name": null,
+                  "ofType": {
+                    "kind": "SCALAR",
+                    "name": "Boolean",
+                    "ofType": null
+                  }
+                }
+              }
+            },
+            "isDeprecated": false,
+            "deprecationReason": null
+          },
+          {
             "name": "upsertValuationRule",
             "description": "",
             "args": [
@@ -13382,6 +13417,31 @@
         "inputFields": [
           {
             "name": "itemCategoriesString",
+            "description": "",
+            "type": {
+              "kind": "NON_NULL",
+              "name": null,
+              "ofType": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              }
+            },
+            "defaultValue": null
+          }
+        ],
+        "interfaces": null,
+        "enumValues": null,
+        "possibleTypes": null
+      },
+      {
+        "kind": "INPUT_OBJECT",
+        "name": "InsertValuationRulesInput",
+        "description": "",
+        "fields": null,
+        "inputFields": [
+          {
+            "name": "valuationRulesString",
             "description": "",
             "type": {
               "kind": "NON_NULL",
