+++ schema.graphql
@@ -112,6 +112,7 @@ type MutationType {
   upsertClaimItem(request: UpsertClaimItemInput): ID!
   deleteClaimItem(claimItemId: ID!): ID
   insertItemCategories(request: InsertItemCategoriesInput): [Boolean!]!
+  insertValuationRules(request: InsertValuationRulesInput): [Boolean!]!
   upsertValuationRule(request: UpsertValuationRuleInput): ID!
   assignCampaignToPartnerPercentageDiscount(
     request: AssignVoucherPercentageDiscount
@@ -119,10 +120,6 @@ type MutationType {
   setContractForClaim(
     request: SetContractForClaim!
   ): Boolean!
-}
-
-input InsertItemCategoriesInput {
-  itemCategoriesString: String!
 }
 
 type ClaimItem {
@@ -136,6 +133,14 @@ type ClaimItem {
   purchasePrice: MonetaryAmountV2
   valuation: MonetaryAmountV2
   note: String
+}
+
+input InsertItemCategoriesInput {
+  itemCategoriesString: String!
+}
+
+input InsertValuationRulesInput {
+  valuationRulesString: String!
 }
 
 input GetValuationInput {
