+++ src/features/tools/itemizer/insert-valuation-rules.graphql
@@ -1,0 +1,3 @@
+mutation InsertValuationRules($request: InsertValuationRulesInput) {
+  insertValuationRules(request: $request)
+}
